+++ compose-app/src/main/java/com/mapbox/maps/compose/testapp/examples/style/StyleStatesActivity.kt
@@ -239,8 +239,9 @@ public class StyleStatesActivity : ComponentActivity() {
                 projection = projections[currentProjectionIdx],
                 atmosphereState = currentAtmosphereState,
                 lightsState = lightsStates[currentLightIdx],
+              ) {
                 lightPreset = lightsPresets[currentLightPresetIdx]
-              )
+              }
             }
           )
         }